requirements.txt
playwright~=1.50.0
loguru
python-dotenv==1.0.0
pyperclip